Tela excluindo avaliação
Tela excluindo avaliação

+++ QuestionarioInfnetEJB/ejbModule/br/edu/infnet/academicnet/modelo/Avaliacao.java
@@ -21,7 +21,7 @@ public class Avaliacao implements Serializable{
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int idAvaliacao;
+	private long idAvaliacao;
 	
 	@Column
 	private String nome;
@@ -63,7 +63,7 @@ public class Avaliacao implements Serializable{
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
-		result = prime * result + idAvaliacao;
+		result = (int) (prime * result + idAvaliacao);
 		result = prime * result
 				+ ((listQuestao == null) ? 0 : listQuestao.hashCode());
 		return result;
@@ -88,11 +88,11 @@ public class Avaliacao implements Serializable{
 		return true;
 	}
 
-	public int getIdAvaliacao() {
+	public long getIdAvaliacao() {
 		return idAvaliacao;
 	}
 
-	public void setIdAvaliacao(int idAvaliacao) {
+	public void setIdAvaliacao(long idAvaliacao) {
 		this.idAvaliacao = idAvaliacao;
 	}
